+++ Funcular.Data.Orm.PostgreSql.Tests/Domain/Entities/Project/ProjectCategoryEntity.cs
@@ -1,0 +1,19 @@
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace Funcular.Data.Orm.PostgreSql.Tests.Domain.Entities.Project
+{
+    [Table("project_category")]
+    public class ProjectCategoryEntity
+    {
+        [Key]
+        [Column("id")]
+        public int Id { get; set; }
+
+        [Column("name")]
+        public string Name { get; set; }
+
+        [Column("code")]
+        public string Code { get; set; }
+    }
+}