+++ Funcular.Data.Orm.SqlServer.Tests/Domain/Entities/Organization/OrganizationEntity.cs
@@ -1,4 +1,5 @@
 using System;
+using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using Funcular.Data.Orm.Attributes;
 using Funcular.Data.Orm.SqlServer.Tests.Domain.Entities.Address;
@@ -13,5 +14,9 @@ namespace Funcular.Data.Orm.SqlServer.Tests.Domain.Entities.Organization
         
         [Column("headquarters_address_id")]
         public int? HeadquartersAddressId { get; set; }
+
+        [Timestamp]
+        [Column("row_version")]
+        public byte[] RowVersion { get; set; }
     }
 }